Added functions updateScore and calculateScore.
Added a data class Cards so I could assign a value to each card.

Wrote the hitButton code to deal cards to the player.

Wrote code for when player push the stand button. Now the dealer will deal cards to themselves.

Added a Handler() so dealers cards are dealt with some delay.

Added some //TODO